requirements.txt
# PDF parsing (PyMuPDF)
pymupdf>=1.23.0

# Web scraping
playwright>=1.40.0
httpx>=0.26.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# CLI & utilities
python-dotenv>=1.0.0
typer>=0.9.0
rich>=13.7.0
tqdm>=4.66.0
